+++ apps/qopnet-admin/src/app/pages/businessOrders/businessOrdersParam.tsx
@@ -1,36 +1,30 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { Link } from 'react-router-dom'
+import { useParams } from 'react-router'
+
 import {
   Box,
   Breadcrumb,
   BreadcrumbItem,
   BreadcrumbLink,
-  Button,
   Flex,
   Heading,
-  Image,
   Spinner,
   Stack,
-  Text,
-  Wrap,
 } from '@chakra-ui/react'
 import { ChevronRightIcon } from '@chakra-ui/icons'
-import { Link } from 'react-router-dom'
-import { useParams, useHistory } from 'react-router'
-import cuid from 'cuid'
 
-import { formatRupiah } from '@qopnet/util-format'
 import { DefaultLayout } from '../../layouts'
 import { useSWR } from '../../utils/swr'
-import { BusinessOrderPaymentSection } from '../../components/businessOrderPaymentSection'
+
+import { BusinessOrderProductSection } from '../../components/businessOrderProductSection'
 import { BusinessOrderStatusSection } from '../../components/businessOrderStatusSection'
+import { BusinessOrderPaymentSection } from '../../components/businessOrderPaymentSection'
 
 export const BusinessOrdersParamPage = () => {
   const { businessOrdersParam }: { businessOrdersParam: string } = useParams()
   const { data, error } = useSWR(`/api/business/orders/${businessOrdersParam}`)
   const { businessOrderParam, businessOrder, message } = data || {}
-
-  const history = useHistory()
-
   return (
     <DefaultLayout>
       <Box px={7} py={7}>
@@ -69,76 +63,8 @@ export const BusinessOrdersParamPage = () => {
               mr={{ lg: 5 }}
             >
               {businessOrder?.businessOrderItems?.length &&
-                businessOrder?.businessOrderItems?.map((item, index) => {
-                  const handleProductDetail = () => {
-                    const supplierName = item?.supplier?.handle
-                    const productSlug = item?.supplierProduct?.slug
-
-                    history.push(
-                      `/suppliers/${supplierName}/products/${productSlug}`
-                    )
-                  }
-
-                  return (
-                    <Wrap
-                      key={cuid()}
-                      id="product-detail"
-                      w="full"
-                      p={5}
-                      spacing={5}
-                      align="center"
-                      border="1px solid"
-                      borderColor="gray.300"
-                      borderRadius="lg"
-                    >
-                      <Image
-                        src={item?.supplierProduct?.images[0]}
-                        boxSize="70"
-                        objectFit="cover"
-                      />
-
-                      <Stack>
-                        <Text fontWeight="bold">
-                          SKU: {item?.supplierProduct?.sku}
-                        </Text>
-                        <Text fontWeight="bold">
-                          {item?.supplierProduct?.name}
-                        </Text>
-
-                        {item?.supplierProduct?.subname && (
-                          <Text fontSize="xs">
-                            {item?.supplierProduct?.subname}
-                          </Text>
-                        )}
-                      </Stack>
-
-                      <Stack align="flex-end" flex="1">
-                        {/* alignSelf="flex-end"{' '} */}
-                        <Text>
-                          {item?.quantity} x{' '}
-                          {formatRupiah(item?.supplierProduct?.price)}
-                        </Text>
-                        <Text fontWeight="bold">
-                          {formatRupiah(
-                            item?.quantity * item?.supplierProduct?.price
-                          )}
-                        </Text>
-                        <Button
-                          bg="orange"
-                          size="xs"
-                          color="white"
-                          onClick={handleProductDetail}
-                          _hover={{
-                            bg: 'none',
-                            border: '1px solid orange',
-                            color: 'orange',
-                          }}
-                        >
-                          Detail Produk
-                        </Button>
-                      </Stack>
-                    </Wrap>
-                  )
+                businessOrder?.businessOrderItems?.map((item) => {
+                  return <BusinessOrderProductSection item={item} />
                 })}
             </Stack>
